+++ vars/terraform.groovy
@@ -137,7 +137,8 @@ void fmt(body) {
     if (config.check == true) {
       cmd += " -check"
     }
-    if (config.write == true) {
+    // incompatible with above
+    else if (config.write == true) {
       cmd += " -write"
     }