+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ### 1.6.0
 **Helm**
 - Drop support for versions < 3.0.
-- Add `plugin` method.
+- Add `plugin` and `status` methods.
 - Add `namespace` parameter to `uninstall` and `rollback` methods.
 
 **Packer**